API Spec: Add ConnectionConfigInfo to ExternalCatalog
Remove the currently unused remoteUrl field from the top-level ExternalCatalog
into the ConnectionConfigInfo as remoteUri instead for better consistency;
remote catalogs in the future may be defined by arbitrary URIs that are not,
for example, http(s) URLs.

This is just the spec definition for now, so it's not yet wired into the
internal entity layer or persistence objects.

Allow extensibility of different connection types in the future even if we
start with only an ICEBERG_REST type.

Similarly, provide extensibility for different authn mechanisms to use
with the connection.