Fix a bug in loadTasks which wasn't abiding by compare-and-swap semantics when updating TaskEntities.
Additionally, fix the PolarisTreeMapMetaStoreSessionImpl to avoid "cheating" by
using its entire PolarisBaseEntit it happens to store in entitiesActive when
performing a listing; in general, entitiesActive is assumed to *only* contain
the name lookup records, so it is not explicitly updated unless a parent or
name changes.

The existing tests happened to pass because of a coincidence of several bugs:

1. Since the entitiesActive and entities happens to share the same PolarisBaseEntity
instance by-reference, any in-place edits to "entities" *happens* to reflect
in entitiesActive, but in general entities *should not be mutable* and usually
a copy of an entity is mutated instead of mutating in place anyways
2. Because of the blind-writes in loadTasks, the "prepare entity" step
where an instance is updated with a new entityVersion coincidentally immediately
causes the main entities *and* entitiesAcitve slices to have the updated entity
even before the call to writeEntity happens
3. Since TreeMap uses "synchronized" for runInTransaction, the loadTasksInParallel
doesn't detect interleaving of threads

So without the fix in the treemap store, simply fixing PolarisMetaStoreManagerImpl
causes the test to run forever because tasks never get updated in entitiesActive
so the same initial 5 tasks keep getting listed forever.

Also fix the test to actually timeout eventually; the Future::get calls were
willing to block forever before getting to the await. Reduce the wait time from
10 minutes to something more reasonable.